Add array support
