+++ membuild/statements.go
@@ -47,6 +47,9 @@ func BuildStmt(p *Program, stmt parser.Statement, instructionum ...int) (Instruc
 	case *parser.RandintStmt:
 		return RandintStmt(p, s)
 
+	case *parser.ArrayStmt:
+		return ArrayStmt(p, s)
+
 	default:
 		return nil, fmt.Errorf("line %d: unknown type %s", stmt.Line(), reflect.TypeOf(stmt).String())
 	}
